PetscMallocA: update macro expansions to cast numbers to (size_t)
Rationale:

  1. Silence -Wconversion warnings.
  2. Your code may be broken because of integer promotion rules.

As an example of (2), imagine that you are running on a 32bit arch with
size_t 32 bits and long long 64 bits. Then

  unsigned long long m1 = 10;
  unsigned long long m2 = 10;
  double *p,*q;
  PetscMalloc1(m1,&p,m2,&q)

Then the macro will compute the product m2*sizeof(double), the result is
then unsigned long long, but PetscMallocA is expecting size_t (because
of va_arg(Argp,size_t)).

Thanks-to: Lisandro Dalcin <dalcinl@gmail.com>

+++ include/petscsys.h
@@ -681,7 +681,7 @@ $  PetscMalloc1(10*sizeof(PetscInt),&id);
   Concepts: memory allocation
 
 M*/
-#define PetscMalloc1(m1,r1) PetscMallocA(1,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1))
+#define PetscMalloc1(m1,r1) PetscMallocA(1,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1))
 
 /*MC
    PetscCalloc1 - Allocates a cleared (zeroed) array of memory aligned to PETSC_MEMALIGN
@@ -707,7 +707,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscCalloc1(m1,r1) PetscMallocA(1,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1))
+#define PetscCalloc1(m1,r1) PetscMallocA(1,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1))
 
 /*MC
    PetscMalloc2 - Allocates 2 arrays of memory both aligned to PETSC_MEMALIGN
@@ -733,7 +733,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscMalloc2(m1,r1,m2,r2) PetscMallocA(2,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2))
+#define PetscMalloc2(m1,r1,m2,r2) PetscMallocA(2,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2))
 
 /*MC
    PetscCalloc2 - Allocates 2 cleared (zeroed) arrays of memory both aligned to PETSC_MEMALIGN
@@ -758,7 +758,7 @@ M*/
 
   Concepts: memory allocation
 M*/
-#define PetscCalloc2(m1,r1,m2,r2) PetscMallocA(2,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2))
+#define PetscCalloc2(m1,r1,m2,r2) PetscMallocA(2,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2))
 
 /*MC
    PetscMalloc3 - Allocates 3 arrays of memory, all aligned to PETSC_MEMALIGN
@@ -786,7 +786,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscMalloc3(m1,r1,m2,r2,m3,r3) PetscMallocA(3,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3))
+#define PetscMalloc3(m1,r1,m2,r2,m3,r3) PetscMallocA(3,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3))
 
 /*MC
    PetscCalloc3 - Allocates 3 cleared (zeroed) arrays of memory, all aligned to PETSC_MEMALIGN
@@ -813,7 +813,7 @@ M*/
 
   Concepts: memory allocation
 M*/
-#define PetscCalloc3(m1,r1,m2,r2,m3,r3) PetscMallocA(3,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3))
+#define PetscCalloc3(m1,r1,m2,r2,m3,r3) PetscMallocA(3,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3))
 
 /*MC
    PetscMalloc4 - Allocates 4 arrays of memory, all aligned to PETSC_MEMALIGN
@@ -843,7 +843,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscMalloc4(m1,r1,m2,r2,m3,r3,m4,r4) PetscMallocA(4,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4))
+#define PetscMalloc4(m1,r1,m2,r2,m3,r3,m4,r4) PetscMallocA(4,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4))
 
 /*MC
    PetscCalloc4 - Allocates 4 cleared (zeroed) arrays of memory, all aligned to PETSC_MEMALIGN
@@ -873,7 +873,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscCalloc4(m1,r1,m2,r2,m3,r3,m4,r4) PetscMallocA(4,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4))
+#define PetscCalloc4(m1,r1,m2,r2,m3,r3,m4,r4) PetscMallocA(4,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4))
 
 /*MC
    PetscMalloc5 - Allocates 5 arrays of memory, all aligned to PETSC_MEMALIGN
@@ -905,7 +905,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscMalloc5(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5) PetscMallocA(5,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4),(m5)*sizeof(**(r5)),(r5))
+#define PetscMalloc5(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5) PetscMallocA(5,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4),(size_t)(m5)*sizeof(**(r5)),(r5))
 
 /*MC
    PetscCalloc5 - Allocates 5 cleared (zeroed) arrays of memory, all aligned to PETSC_MEMALIGN
@@ -937,7 +937,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscCalloc5(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5) PetscMallocA(5,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4),(m5)*sizeof(**(r5)),(r5))
+#define PetscCalloc5(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5) PetscMallocA(5,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4),(size_t)(m5)*sizeof(**(r5)),(r5))
 
 /*MC
    PetscMalloc6 - Allocates 6 arrays of memory, all aligned to PETSC_MEMALIGN
@@ -971,7 +971,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscMalloc6(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6) PetscMallocA(6,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4),(m5)*sizeof(**(r5)),(r5),(m6)*sizeof(**(r6)),(r6))
+#define PetscMalloc6(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6) PetscMallocA(6,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4),(size_t)(m5)*sizeof(**(r5)),(r5),(size_t)(m6)*sizeof(**(r6)),(r6))
 
 /*MC
    PetscCalloc6 - Allocates 6 cleared (zeroed) arrays of memory, all aligned to PETSC_MEMALIGN
@@ -1004,7 +1004,7 @@ M*/
 
   Concepts: memory allocation
 M*/
-#define PetscCalloc6(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6) PetscMallocA(6,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4),(m5)*sizeof(**(r5)),(r5),(m6)*sizeof(**(r6)),(r6))
+#define PetscCalloc6(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6) PetscMallocA(6,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4),(size_t)(m5)*sizeof(**(r5)),(r5),(size_t)(m6)*sizeof(**(r6)),(r6))
 
 /*MC
    PetscMalloc7 - Allocates 7 arrays of memory, all aligned to PETSC_MEMALIGN
@@ -1040,7 +1040,7 @@ M*/
   Concepts: memory allocation
 
 M*/
-#define PetscMalloc7(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6,m7,r7) PetscMallocA(7,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4),(m5)*sizeof(**(r5)),(r5),(m6)*sizeof(**(r6)),(r6),(m7)*sizeof(**(r7)),(r7))
+#define PetscMalloc7(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6,m7,r7) PetscMallocA(7,PETSC_FALSE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4),(size_t)(m5)*sizeof(**(r5)),(r5),(size_t)(m6)*sizeof(**(r6)),(r6),(size_t)(m7)*sizeof(**(r7)),(r7))
 
 /*MC
    PetscCalloc7 - Allocates 7 cleared (zeroed) arrays of memory, all aligned to PETSC_MEMALIGN
@@ -1075,7 +1075,7 @@ M*/
 
   Concepts: memory allocation
 M*/
-#define PetscCalloc7(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6,m7,r7) PetscMallocA(7,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(m1)*sizeof(**(r1)),(r1),(m2)*sizeof(**(r2)),(r2),(m3)*sizeof(**(r3)),(r3),(m4)*sizeof(**(r4)),(r4),(m5)*sizeof(**(r5)),(r5),(m6)*sizeof(**(r6)),(r6),(m7)*sizeof(**(r7)),(r7))
+#define PetscCalloc7(m1,r1,m2,r2,m3,r3,m4,r4,m5,r5,m6,r6,m7,r7) PetscMallocA(7,PETSC_TRUE,__LINE__,PETSC_FUNCTION_NAME,__FILE__,(size_t)(m1)*sizeof(**(r1)),(r1),(size_t)(m2)*sizeof(**(r2)),(r2),(size_t)(m3)*sizeof(**(r3)),(r3),(size_t)(m4)*sizeof(**(r4)),(r4),(size_t)(m5)*sizeof(**(r5)),(r5),(size_t)(m6)*sizeof(**(r6)),(r6),(size_t)(m7)*sizeof(**(r7)),(r7))
 
 /*MC
    PetscNew - Allocates memory of a particular type, zeros the memory! Aligned to PETSC_MEMALIGN
